+++ routes/seed.js
@@ -1,58 +1,47 @@
-const
-    express                 = require('express'),
-    router                  = express.Router();
+const RouterHub = require('../controllers/router');
 
 
 
 // ==============================
-//  MODELS
+//  CONTROLLER
 // ==============================
-const { postModel, userModel } = require('../models');
-
-
-
-// ==============================
-//  FUNCTIONS
-// ==============================
-// middleware
-const { _md } = require('../controllers/middleware');
-
-
-
-// ==============================
-//  ROUTE RULES
-// ==============================
-// seed data
-const SEEDUSER = {
-    email       : 'leo@leoyli.com',
-    username    : 'leo',
-    password    : 'leo',
-    firstName   : 'Leo',
-    lastName    : 'Li',
-    picture     : '/media/201710/1509304639065.png'
-};
-
-const SEEDPOST = {
-    title       : 'New weapon arrived: Custom E-liter 4K!',
-    author      : { _id: SEEDUSER._id, username: SEEDUSER.username },
-    featured    : 'http://www.perfectly-nintendo.com/wp-content/gallery/splatoon-2-13-10-2017/1.jpg',
-    content     : 'New arrived: <strong>Custom E-liter 4K!</strong><script>alert("WARNING");</script>',
+const moc = {
+    user: {
+        email       : 'leo@leoyli.com',
+        username    : 'leo',
+        password    : 'leo',
+        firstName   : 'Leo',
+        lastName    : 'Li',
+        picture     : '/media/201710/1509304639065.png'
+    },
+    post : function(newUser) { return {
+        title       : 'New weapon arrived: Custom E-liter 4K!',
+        author      : { _id: newUser._id, username: newUser.username },
+        featured    : 'http://www.perfectly-nintendo.com/wp-content/gallery/splatoon-2-13-10-2017/1.jpg',
+        content     : 'New arrived: <strong>Custom E-liter 4K!</strong><script>alert("WARNING");</script>',
+    }}
 };
 
 
-// seed plant
-router.get("/", _md.wrapAsync(async (req, res) => {
-    const newUser = await userModel.register(new userModel(SEEDUSER), SEEDUSER.password);
-    await postModel.postsCreateThenAssociate(SEEDPOST, newUser);
+const seed = require('../controllers/middleware')._md.wrapAsync(async (req, res) => {
+    const { postModel, userModel } = require('../models');
+    const newUser = await userModel.register(new userModel(moc.user), moc.user.password);
+    await postModel.postsCreateThenAssociate(moc.post(newUser), newUser);
     req.flash('info', 'Successfully seeded.');
     res.redirect('/post');
-}));
-
-
-// error handler
-router.use(require('../controllers/render').errorHandler);
+});
 
 
 
-// route exports
-module.exports  = router;
+// ==============================
+//  ROUTER HUB
+// ==============================
+const SeedRouter = new RouterHub([{
+    route: '/',
+    controller: seed,
+}]);
+
+
+
+// router exports
+module.exports = SeedRouter.activate();